+++ packages/db/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.29.3](https://github.com/proteinjs/db/compare/@proteinjs/db@1.29.2...@proteinjs/db@1.29.3) (2026-08-14)
+
+**Note:** Version bump only for package @proteinjs/db
+
+
+
+
+
+## [1.29.2](https://github.com/proteinjs/db/compare/@proteinjs/db@1.29.1...@proteinjs/db@1.29.2) (2026-08-13)
+
+**Note:** Version bump only for package @proteinjs/db
+
+
+
+
+
 ## [1.29.1](https://github.com/proteinjs/db/compare/@proteinjs/db@1.29.0...@proteinjs/db@1.29.1) (2026-08-13)