+++ packages/spanner-common/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.5](https://github.com/proteinjs/db/compare/@proteinjs/db-spanner-common@1.3.4...@proteinjs/db-spanner-common@1.3.5) (2026-08-14)
+
+**Note:** Version bump only for package @proteinjs/db-spanner-common
+
+
+
+
+
+## [1.3.4](https://github.com/proteinjs/db/compare/@proteinjs/db-spanner-common@1.3.3...@proteinjs/db-spanner-common@1.3.4) (2026-08-13)
+
+**Note:** Version bump only for package @proteinjs/db-spanner-common
+
+
+
+
+
 ## [1.3.3](https://github.com/proteinjs/db/compare/@proteinjs/db-spanner-common@1.3.2...@proteinjs/db-spanner-common@1.3.3) (2026-08-13)
 
 **Note:** Version bump only for package @proteinjs/db-spanner-common